fix: FILES-302 - Fix postgresql dependency on RHEL (#1)
In debian based distro the psql tool is contained in the
postgresql-client package. In RHEL based distro the tool is contained in
the postgresql package. The PKBGUILD was changed accordingly.

A minor changed was made in the Jenkinsfile in order to replace the
pacur build target from centos8 to rocky8.